Create DrainOddOrEvenOctopusServerNodesForDifferingHostnames.ps1
One of my customers needs to drain alternate nodes in order for them to be patched however, it's a 6-node HA cluster with the 1st two nodes being UI-only nodes and named differently to the other four. PLUS, there are 4 trailing numbers in the node's hostnames.

This variation on the original script tries to take that all into account.